refactor: remove unused test classes and update package structure
