--- citizen-intelligence-agency/src/test/java/com/hack23/cia/systemintegrationtest/AbstractRoleSystemTest.java
@@ -1,14 +1,0 @@
-package com.hack23.cia.systemintegrationtest;
-
-public abstract class AbstractRoleSystemTest extends AbstractSystemIntegrationTest {
-    protected final Browser browser;
-
-    protected AbstractRoleSystemTest(final Browser browser) {
-        this.browser = browser;
-    }
-
-    @Override
-    protected Browser getBrowser() {
-        return browser;
-    }
-}
