+++ citizen-intelligence-agency/src/test/java/com/hack23/cia/systemintegrationtest/AdminDataTest.java
@@ -7,21 +7,20 @@ import org.junit.runners.Parameterized;
 
 import com.hack23.cia.systemintegrationtest.suites.IntegrationTest;
 import com.hack23.cia.web.impl.ui.application.views.common.pagelinks.api.PageModeMenuCommand;
-import com.hack23.cia.web.impl.ui.application.views.common.viewnames.UserViews;
+import com.hack23.cia.web.impl.ui.application.views.common.viewnames.AdminViews;
 
 @Category(IntegrationTest.class)
-public final class BallotViewTest extends AbstractUITest {
+public final class AdminDataTest extends AbstractUITest {
 
- 
-    @Test(timeout = DEFAULT_TIMEOUT)
-    public void testBallotView() throws Exception {
+    @Test(timeout = TestConstants.DEFAULT_TIMEOUT)
+    public void shouldViewAdminData() throws Exception {
         retryOnFailure(() -> {
             try {
-                pageVisit.visitDirectPage(new PageModeMenuCommand(UserViews.BALLOT_VIEW_NAME, ""));
-                pageVisit.verifyPageContent("Ballot View");
+                pageVisit.visitDirectPage(new PageModeMenuCommand(AdminViews.ADMIN_DATA_SUMMARY_VIEW_NAME, ""));
+                pageVisit.verifyPageContent("Admin Data Summary");
             } catch (Exception e) {
                 throw new RuntimeException(e);
             }
-        }, 3);
+        }, TestConstants.DEFAULT_MAX_RETRIES);
     }
 }
