+++ citizen-intelligence-agency/src/test/java/com/hack23/cia/systemintegrationtest/AdminMonitoringTest.java
@@ -1,8 +1,9 @@
-package com.hack23.cia.systemintegrationtest.suites;
+package com.hack23.cia.systemintegrationtest;
 
 import org.junit.Test;
 import org.junit.experimental.categories.Category;
-import com.hack23.cia.systemintegrationtest.IntegrationTest;
+
+import com.hack23.cia.systemintegrationtest.suites.IntegrationTest;
 
 @Category(IntegrationTest.class)
 public class AdminMonitoringTest {
