+++ citizen-intelligence-agency/src/test/java/com/hack23/cia/systemintegrationtest/AdminRoleSystemITest.java
@@ -18,11 +18,15 @@
 */
 package com.hack23.cia.systemintegrationtest;
 
-import static com.hack23.cia.systemintegrationtest.TestConstants.*;
+import static com.hack23.cia.systemintegrationtest.TestConstants.DEFAULT_MAX_RETRIES;
+
+import org.junit.Test;
 import org.junit.experimental.categories.Category;
 import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
 import org.openqa.selenium.WebElement;
+
+import com.hack23.cia.systemintegrationtest.suites.IntegrationTest;
 
 // import com.hack23.cia.systemintegrationtest.TestConstants; // Unused import removed
 // import com.hack23.cia.web.impl.ui.application.action.ViewAction; // Unused import removed
@@ -32,36 +36,15 @@ import com.hack23.cia.web.impl.ui.application.views.common.viewnames.AdminViews;
 import com.hack23.cia.web.impl.ui.application.views.common.viewnames.DataSummaryPageMode;
 import com.hack23.cia.web.impl.ui.application.views.common.viewnames.PageMode;
 
-import org.junit.Test;
-import org.junit.runners.Parameterized.Parameters;
-import java.util.Arrays;
-import java.util.Collection;
-
-@RunWith(Parameterized.class)
 @Category(IntegrationTest.class)
 public final class AdminRoleSystemITest extends AbstractUITest {
 
-    private final Browser browser;
-
-    public AdminRoleSystemITest(final Browser browser) {
-        this.browser = browser;
-    }
-
-    @Override
-    protected Browser getBrowser() {
-        return browser;
-    }
-
-    @Parameters(name = "AdminRoleSiteTest{index}: browser({0})")
-    public static Collection<Browser[]> browsers() {
-        return Arrays.asList(new Browser[][] { { Browser.CHROME } });
-    }
 
     @Test(timeout = DEFAULT_TIMEOUT)
     public void shouldAccessAdminAgencyPage() {
         retryOnFailure(() -> {
             try {
-                loginAsAdmin(pageVisit);
+                pageVisit.loginAsAdmin(pageVisit);
                 pageVisit.visitDirectPage(new PageModeMenuCommand(AdminViews.ADMIN_AGENCY_VIEW_NAME, ""));
                 pageVisit.verifyPageContent("Agency");
                 pageVisit.selectFirstGridRow();
@@ -77,7 +60,7 @@ public final class AdminRoleSystemITest extends AbstractUITest {
         try {
             retryOnFailure(() -> {
                 try {
-                    loginAsAdmin(pageVisit);
+                    pageVisit.loginAsAdmin(pageVisit);
                     pageVisit.visitDirectPage(new PageModeMenuCommand(AdminViews.ADMIN_APPLICATIONS_EVENTS_VIEW_NAME, PageMode.CHARTS));
                     pageVisit.verifyPageContent("Charts");
                     pageVisit.validatePage(new PageModeMenuCommand(AdminViews.ADMIN_APPLICATIONS_EVENTS_VIEW_NAME, PageMode.CHARTS));
@@ -95,7 +78,7 @@ public final class AdminRoleSystemITest extends AbstractUITest {
         try {
             retryOnFailure(() -> {
                 try {
-                    loginAsAdmin(pageVisit);
+                    pageVisit.loginAsAdmin(pageVisit);
                     pageVisit.visitDirectPage(new PageModeMenuCommand(AdminViews.ADMIN_APPLICATIONS_EVENTS_VIEW_NAME, ""));
                     pageVisit.verifyPageContent("Admin Application Events");
                     pageVisit.selectFirstGridRow();
@@ -114,7 +97,7 @@ public final class AdminRoleSystemITest extends AbstractUITest {
         try {
             retryOnFailure(() -> {
                 try {
-                    loginAsAdmin(pageVisit);
+                    pageVisit.loginAsAdmin(pageVisit);
                     pageVisit.visitDirectPage(new PageModeMenuCommand(AdminViews.ADMIN_APPLICATIONS_SESSION_VIEW_NAME, PageMode.CHARTS));
                     pageVisit.verifyPageContent("Admin Application Session Charts");
                     pageVisit.validatePage(new PageModeMenuCommand(AdminViews.ADMIN_APPLICATIONS_SESSION_VIEW_NAME, PageMode.CHARTS));
@@ -132,7 +115,7 @@ public final class AdminRoleSystemITest extends AbstractUITest {
         try {
             retryOnFailure(() -> {
                 try {
-                    loginAsAdmin(pageVisit);
+                    pageVisit.loginAsAdmin(pageVisit);
                     pageVisit.visitDirectPage(new PageModeMenuCommand(AdminViews.ADMIN_APPLICATIONS_SESSION_VIEW_NAME, ""));
                     pageVisit.verifyPageContent("Application Session");
                     final WebElement nextPageButton = pageVisit.findButton("last page");
@@ -151,7 +134,7 @@ public final class AdminRoleSystemITest extends AbstractUITest {
         try {
             retryOnFailure(() -> {
                 try {
-                    loginAsAdmin(pageVisit);
+                    pageVisit.loginAsAdmin(pageVisit);
                     pageVisit.visitDirectPage(new PageModeMenuCommand(AdminViews.ADMIN_APPLICATIONS_SESSION_VIEW_NAME, ""));
                     pageVisit.verifyPageContent("Application Session");
                     final WebElement nextPageButton = pageVisit.findButton("next page");
@@ -170,7 +153,7 @@ public final class AdminRoleSystemITest extends AbstractUITest {
         try {
             retryOnFailure(() -> {
                 try {
-                    loginAsAdmin(pageVisit);
+                    pageVisit.loginAsAdmin(pageVisit);
                     pageVisit.visitDirectPage(new PageModeMenuCommand(AdminViews.ADMIN_APPLICATIONS_SESSION_VIEW_NAME, ""));
                     pageVisit.verifyPageContent("Application Session");
                     pageVisit.selectFirstGridRow();
@@ -206,7 +189,7 @@ public final class AdminRoleSystemITest extends AbstractUITest {
         try {
             retryOnFailure(() -> {
                 try {
-                    loginAsAdmin(pageVisit);
+                    pageVisit.loginAsAdmin(pageVisit);
                     pageVisit.visitDirectPage(new PageModeMenuCommand(AdminViews.ADMIN_MONITORING_VIEW_NAME, ""));
                     pageVisit.verifyPageContent("Admin Monitoring");
                     assertFalse("Dont expect this content", pageVisit.getIframesHtmlBodyAsText().contains("Login with Username and Password"));
@@ -224,7 +207,7 @@ public final class AdminRoleSystemITest extends AbstractUITest {
         try {
             retryOnFailure(() -> {
                 try {
-                    loginAsAdmin(pageVisit);
+                    pageVisit.loginAsAdmin(pageVisit);
                     pageVisit.visitDirectPage(new PageModeMenuCommand(AdminViews.ADMIN_PORTAL_VIEW_NAME, ""));
                     pageVisit.verifyPageContent("Portal");
                     pageVisit.selectFirstGridRow();
@@ -243,7 +226,7 @@ public final class AdminRoleSystemITest extends AbstractUITest {
         try {
             retryOnFailure(() -> {
                 try {
-                    loginAsAdmin(pageVisit);
+                    pageVisit.loginAsAdmin(pageVisit);
                     pageVisit.visitDirectPage(new PageModeMenuCommand(AdminViews.ADMIN_DATA_SUMMARY_VIEW_NAME, DataSummaryPageMode.AUTHORS.toString()));
                 } catch (Exception e) {
                     throw new RuntimeException(e);
@@ -259,7 +242,7 @@ public final class AdminRoleSystemITest extends AbstractUITest {
         try {
             retryOnFailure(() -> {
                 try {
-                    loginAsAdmin(pageVisit);
+                    pageVisit.loginAsAdmin(pageVisit);
                     pageVisit.visitDirectPage(new PageModeMenuCommand(AdminViews.ADMIN_DATA_SUMMARY_VIEW_NAME, ""));
                     final WebElement refreshViewsButton = pageVisit.findButton("Refresh Views");
                     assertNotNull("Expect to find a Refresh Views Button", refreshViewsButton);
@@ -278,7 +261,7 @@ public final class AdminRoleSystemITest extends AbstractUITest {
         try {
             retryOnFailure(() -> {
                 try {
-                    loginAsAdmin(pageVisit);
+                    pageVisit.loginAsAdmin(pageVisit);
                     pageVisit.visitDirectPage(new PageModeMenuCommand(AdminViews.ADMIN_DATA_SUMMARY_VIEW_NAME, ""));
                     final WebElement removeApplicationHistoryButton = pageVisit.findButton("Remove Application History");
                     assertNotNull("Expect to find a Button", removeApplicationHistoryButton);
@@ -296,7 +279,7 @@ public final class AdminRoleSystemITest extends AbstractUITest {
         try {
             retryOnFailure(() -> {
                 try {
-                    loginAsAdmin(pageVisit);
+                    pageVisit.loginAsAdmin(pageVisit);
                     pageVisit.visitDirectPage(new PageModeMenuCommand(AdminViews.ADMIN_DATA_SUMMARY_VIEW_NAME, ""));
                     final WebElement removeDocumentsButton = pageVisit.findButton("Remove Documents");
                     assertNotNull("Expect to find a Button", removeDocumentsButton);
@@ -314,7 +297,7 @@ public final class AdminRoleSystemITest extends AbstractUITest {
         try {
             retryOnFailure(() -> {
                 try {
-                    loginAsAdmin(pageVisit);
+                    pageVisit.loginAsAdmin(pageVisit);
                     pageVisit.visitDirectPage(new PageModeMenuCommand(AdminViews.ADMIN_DATA_SUMMARY_VIEW_NAME, ""));
                     final WebElement removePoliticiansButton = pageVisit.findButton("Remove Politicians");
                     assertNotNull("Expect to find a Button", removePoliticiansButton);
@@ -332,7 +315,7 @@ public final class AdminRoleSystemITest extends AbstractUITest {
         try {
             retryOnFailure(() -> {
                 try {
-                    loginAsAdmin(pageVisit);
+                    pageVisit.loginAsAdmin(pageVisit);
                     pageVisit.visitDirectPage(new PageModeMenuCommand(AdminViews.ADMIN_DATA_SUMMARY_VIEW_NAME, ""));
                 } catch (Exception e) {
                     throw new RuntimeException(e);
@@ -348,7 +331,7 @@ public final class AdminRoleSystemITest extends AbstractUITest {
         try {
             retryOnFailure(() -> {
                 try {
-                    loginAsAdmin(pageVisit);
+                    pageVisit.loginAsAdmin(pageVisit);
                     pageVisit.visitDirectPage(new PageModeMenuCommand(AdminViews.ADMIN_DATA_SUMMARY_VIEW_NAME, ""));
                     final WebElement updateSearchIndexButton = pageVisit.findButton("Update Search Index");
                     assertNotNull("Expect to find a Update Search Index Button", updateSearchIndexButton);
