--- citizen-intelligence-agency/src/test/java/com/hack23/cia/systemintegrationtest/Browser.java
@@ -1,7 +1,0 @@
-package com.hack23.cia.systemintegrationtest;
-
-public enum Browser {
-    CHROME,
-    FIREFOX,
-    EDGE;
-}
