--- citizen-intelligence-agency/src/test/java/com/hack23/cia/systemintegrationtest/CommonTestImports.java
@@ -1,5 +1,0 @@
-package com.hack23.cia.systemintegrationtest;
-
-public abstract class CommonTestImports {
-    public static final long DEFAULT_TIMEOUT = 60000;
-}
