+++ citizen-intelligence-agency/src/test/java/com/hack23/cia/systemintegrationtest/DocumentViewTest.java
@@ -10,15 +10,17 @@ import com.hack23.cia.web.impl.ui.application.views.common.pagelinks.api.PageMod
 import com.hack23.cia.web.impl.ui.application.views.common.viewnames.UserViews;
 
 @Category(IntegrationTest.class)
-public final class BallotViewTest extends AbstractUITest {
+public final class DocumentViewTest extends AbstractUITest {
 
- 
+
     @Test(timeout = DEFAULT_TIMEOUT)
-    public void testBallotView() throws Exception {
+    public void testDocumentView() throws Exception {
         retryOnFailure(() -> {
             try {
-                pageVisit.visitDirectPage(new PageModeMenuCommand(UserViews.BALLOT_VIEW_NAME, ""));
-                pageVisit.verifyPageContent("Ballot View");
+                pageVisit.visitDirectPage(new PageModeMenuCommand(UserViews.DOCUMENT_VIEW_NAME, ""));
+                pageVisit.verifyPageContent("Document View");
+                pageVisit.selectFirstGridRow();
+                pageVisit.validatePage(new PageModeMenuCommand(UserViews.DOCUMENT_VIEW_NAME, ""));
             } catch (Exception e) {
                 throw new RuntimeException(e);
             }
