+++ citizen-intelligence-agency/src/test/java/com/hack23/cia/systemintegrationtest/MinistryViewTest.java
@@ -1,20 +1,18 @@
 package com.hack23.cia.systemintegrationtest;
 
+import static com.hack23.cia.systemintegrationtest.TestConstants.DEFAULT_MAX_RETRIES;
+
 import org.junit.Test;
 import org.junit.experimental.categories.Category;
 
+import com.hack23.cia.systemintegrationtest.suites.IntegrationTest;
 import com.hack23.cia.web.impl.ui.application.views.common.pagelinks.api.PageModeMenuCommand;
 import com.hack23.cia.web.impl.ui.application.views.common.viewnames.PageMode;
 import com.hack23.cia.web.impl.ui.application.views.common.viewnames.UserViews;
-import static com.hack23.cia.systemintegrationtest.TestConstants.*;
 
 @Category(IntegrationTest.class)
 public class MinistryViewTest extends AbstractUITest {
     
-    @Override
-    protected Browser getBrowser() {
-        return Browser.CHROME;
-    }
 
     @Test(timeout = DEFAULT_TIMEOUT)
     public void testMinistryView() throws Exception {
