+++ citizen-intelligence-agency/src/test/java/com/hack23/cia/systemintegrationtest/PartyViewTest.java
@@ -3,6 +3,7 @@ package com.hack23.cia.systemintegrationtest;
 import org.junit.Test;
 import org.junit.experimental.categories.Category;
 
+import com.hack23.cia.systemintegrationtest.suites.IntegrationTest;
 import com.hack23.cia.web.impl.ui.application.views.common.pagelinks.api.PageModeMenuCommand;
 import com.hack23.cia.web.impl.ui.application.views.common.viewnames.PartyPageMode;
 import com.hack23.cia.web.impl.ui.application.views.common.viewnames.UserViews;
@@ -10,18 +11,8 @@ import com.hack23.cia.web.impl.ui.application.views.common.viewnames.UserViews;
 @Category(IntegrationTest.class)
 public final class PartyViewTest extends AbstractUITest {
     private static final int MAX_RETRIES = 3;
-    private final Browser browser;
-
-    public PartyViewTest(final Browser browser) {
-        this.browser = browser;
-    }
-
-    @Override
-    protected Browser getBrowser() {
-        return browser;
-    }
-
-    @Test(timeout = 60000)
+ 
+    @Test(timeout = DEFAULT_TIMEOUT)
     public void shouldViewPartyCommitteeRoles() throws Exception {
         retryOnFailure(() -> {
             try {
@@ -34,7 +25,7 @@ public final class PartyViewTest extends AbstractUITest {
         }, MAX_RETRIES);
     }
 
-    @Test(timeout = 60000)
+    @Test(timeout = DEFAULT_TIMEOUT)
     public void shouldViewPartyDocumentHistory() throws Exception {
         retryOnFailure(() -> {
             try {
