+++ citizen-intelligence-agency/src/test/java/com/hack23/cia/systemintegrationtest/PoliticianViewTest.java
@@ -1,23 +1,17 @@
-package com.hack23.cia.systemintegrationtest.suites;
+package com.hack23.cia.systemintegrationtest;
 
 import org.junit.Test;
 import org.junit.experimental.categories.Category;
 
-import com.hack23.cia.systemintegrationtest.AbstractUITest;
-import com.hack23.cia.systemintegrationtest.Browser;
-import com.hack23.cia.systemintegrationtest.IntegrationTest;
+import com.hack23.cia.systemintegrationtest.suites.IntegrationTest;
 import com.hack23.cia.web.impl.ui.application.views.common.pagelinks.api.PageModeMenuCommand;
 import com.hack23.cia.web.impl.ui.application.views.common.viewnames.UserViews;
 
 @Category(IntegrationTest.class)
 public class PoliticianViewTest extends AbstractUITest {
     
-    @Override
-    protected Browser getBrowser() {
-        return Browser.CHROME;
-    }
 
-    @Test(timeout = 60000)
+    @Test(timeout = DEFAULT_TIMEOUT)
     public void testPoliticianView() throws Exception {
         retryOnFailure(() -> {
             try {
