+++ citizen-intelligence-agency/src/test/java/com/hack23/cia/systemintegrationtest/UserNavigationTest.java
@@ -1,20 +1,17 @@
 package com.hack23.cia.systemintegrationtest;
 
-import java.util.Arrays;
-import java.util.Collection;
 import java.util.List;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
-import org.junit.runners.Parameterized.Parameters;
 
 import com.hack23.cia.web.impl.ui.application.action.ViewAction;
 import com.hack23.cia.web.impl.ui.application.views.common.pagelinks.api.PageModeMenuCommand;
 import com.hack23.cia.web.impl.ui.application.views.common.viewnames.PageMode;
 import com.hack23.cia.web.impl.ui.application.views.common.viewnames.UserViews;
 
-@RunWith(Parameterized.class)
+@Category(IntegrationTest.class)
 public final class UserNavigationTest extends AbstractUITest {
     private static final int MAX_RETRIES = 3;
     // Constants for page modes
@@ -23,23 +20,7 @@ public final class UserNavigationTest extends AbstractUITest {
     private static final String DECISION_FLOW_CHART = "DECISION_FLOW_CHART";
     private static final String RISK_SUMMARY = "RISK_SUMMARY";
 
-    private final Browser browser;
-
-    public UserNavigationTest(final Browser browser) {
-        this.browser = browser;
-    }
-
-    @Override
-    protected Browser getBrowser() {
-        return browser;
-    }
-
-    @Parameters(name = "UserNavTest{index}: browser({0})")
-    public static Collection<Browser[]> browsers() {
-        return Arrays.asList(new Browser[][] { { Browser.CHROME } });
-    }
-
-    @Test(timeout = 60000)
+    @Test(timeout = DEFAULT_TIMEOUT)
     public void shouldNavigatePoliticianViews() throws Exception {
         retryOnFailure(() -> {
             try {
@@ -65,7 +46,7 @@ public final class UserNavigationTest extends AbstractUITest {
         }, MAX_RETRIES);
     }
 
-    @Test(timeout = 60000)
+    @Test(timeout = DEFAULT_TIMEOUT)
     public void shouldNavigateParliamentViews() throws Exception {
         retryOnFailure(() -> {
             try {
@@ -89,7 +70,7 @@ public final class UserNavigationTest extends AbstractUITest {
         }, MAX_RETRIES);
     }
 
-    @Test(timeout = 60000)
+    @Test(timeout = DEFAULT_TIMEOUT)
     public void shouldNavigateCommitteeRanking() throws Exception {
         retryOnFailure(() -> {
             try {
@@ -101,7 +82,7 @@ public final class UserNavigationTest extends AbstractUITest {
         }, MAX_RETRIES);
     }
 
-    @Test(timeout = 60000)
+    @Test(timeout = DEFAULT_TIMEOUT)
     public void shouldNavigateMinistryRanking() throws Exception {
         retryOnFailure(() -> {
             try {
@@ -113,7 +94,7 @@ public final class UserNavigationTest extends AbstractUITest {
         }, MAX_RETRIES);
     }
 
-    @Test(timeout = 60000)
+    @Test(timeout = DEFAULT_TIMEOUT)
     public void shouldNavigatePartyRanking() throws Exception {
         retryOnFailure(() -> {
             try {
@@ -128,7 +109,7 @@ public final class UserNavigationTest extends AbstractUITest {
         }, MAX_RETRIES);
     }
 
-    @Test(timeout = 60000)
+    @Test(timeout = DEFAULT_TIMEOUT)
     public void shouldNavigatePoliticianRanking() throws Exception {
         retryOnFailure(() -> {
             try {
