+++ citizen-intelligence-agency/src/test/java/com/hack23/cia/systemintegrationtest/UserRoleSystemITest.java
@@ -18,41 +18,20 @@
 */
 package com.hack23.cia.systemintegrationtest;
 
-import java.util.Arrays;
-import java.util.Collection;
-
 import org.junit.Test;
 import org.junit.experimental.categories.Category;
 import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
-import org.junit.runners.Parameterized.Parameters;
 
+import com.hack23.cia.systemintegrationtest.suites.IntegrationTest;
 import com.hack23.cia.web.impl.ui.application.views.common.pagelinks.api.PageModeMenuCommand;
 import com.hack23.cia.web.impl.ui.application.views.common.viewnames.PageMode;
 import com.hack23.cia.web.impl.ui.application.views.common.viewnames.UserHomePageMode;
 import com.hack23.cia.web.impl.ui.application.views.common.viewnames.UserViews;
 
-@RunWith(Parameterized.class)
 @Category(IntegrationTest.class)
 public final class UserRoleSystemITest extends AbstractUITest {
 
-    private final Browser browser;
-    private final UserPageVisit pageVisit;
-
-    public UserRoleSystemITest(final Browser browser) {
-        this.browser = browser;
-        this.pageVisit = new UserPageVisit(getWebDriver(), browser);
-    }
-
-    @Override
-    protected Browser getBrowser() {
-        return browser;
-    }
-
-    @Parameters(name = "UserRoleSiteTest{index}: browser({0})")
-    public static Collection<Browser[]> browsers() {
-        return Arrays.asList(new Browser[][] { { Browser.CHROME } });
-    }
 
     @Test(timeout = DEFAULT_TIMEOUT)
     public void shouldViewMinistryContent() throws Exception {
