+++ citizen-intelligence-agency/src/test/java/com/hack23/cia/systemintegrationtest/suites/IntegrationTest.java
@@ -1,4 +1,4 @@
-package com.hack23.cia.systemintegrationtest;
+package com.hack23.cia.systemintegrationtest.suites;
 
 public interface IntegrationTest {
     // Marker interface for integration tests
