+++ citizen-intelligence-agency/src/test/java/com/hack23/cia/systemintegrationtest/suites/IntegrationTestSuite.java
@@ -5,14 +5,21 @@ import org.junit.runners.Suite;
 
 import com.hack23.cia.systemintegrationtest.AdminAgentTest;
 import com.hack23.cia.systemintegrationtest.AdminConfigurationTest;
+import com.hack23.cia.systemintegrationtest.AdminDataTest;
 import com.hack23.cia.systemintegrationtest.AdminEmailTest;
 import com.hack23.cia.systemintegrationtest.AdminUserManagementTest;
 import com.hack23.cia.systemintegrationtest.AuthorizationTest;
+import com.hack23.cia.systemintegrationtest.BallotViewTest;
+import com.hack23.cia.systemintegrationtest.CommitteeViewTest;
+import com.hack23.cia.systemintegrationtest.MinistryViewTest;
 import com.hack23.cia.systemintegrationtest.PartyViewTest;
+import com.hack23.cia.systemintegrationtest.PoliticianViewTest;
+import com.hack23.cia.systemintegrationtest.SearchViewTest;
 import com.hack23.cia.systemintegrationtest.SessionManagementTest;
 import com.hack23.cia.systemintegrationtest.UserAuthenticationTest;
 import com.hack23.cia.systemintegrationtest.UserNavigationTest;
 import com.hack23.cia.systemintegrationtest.ViewPageTest;
+import com.hack23.cia.systemintegrationtest.ViewRankingTest;
 
 @RunWith(Suite.class)
 @Suite.SuiteClasses({
@@ -23,7 +30,6 @@ import com.hack23.cia.systemintegrationtest.ViewPageTest;
         PartyViewTest.class,
         MinistryViewTest.class,
         PoliticianViewTest.class,
-        DocumentViewTest.class,
         CommitteeViewTest.class,
         BallotViewTest.class,
 
@@ -35,7 +41,6 @@ import com.hack23.cia.systemintegrationtest.ViewPageTest;
         // Admin Tests
         AdminConfigurationTest.class,
         AdminUserManagementTest.class,
-        AdminMonitoringTest.class,
         AdminDataTest.class,
         AdminEmailTest.class,
         AdminAgentTest.class,
