+++ service.external.worldbank/src/main/java/module-info.java
@@ -87,6 +87,9 @@ open module com.hack23.cia.service.external.worldbank {
 	requires jakarta.activation;
 	requires org.apache.commons.lang3;
 	requires org.apache.commons.csv;
+	requires org.apache.httpcomponents.httpclient;
+    requires org.apache.httpcomponents.httpclient.fluent;
+
 
 	requires transitive com.hack23.cia.service.external.common;
 	requires transitive com.hack23.cia.model.external.worldbank.countries.impl;
